requirements.txt
streamlit
requests
beautifulsoup4
openai
python-dotenv
pypdf
python-docx